+++ plugin.d.ts
@@ -1,0 +1,6 @@
+import type { Config, PluginCreator } from './types/config'
+declare function createPlugin(
+  plugin: PluginCreator,
+  config?: Config
+): { handler: PluginCreator; config?: Config }
+export = createPlugin